Changed type of input from int to double

+++ src/application/HeptaPrismController.java
@@ -41,8 +41,8 @@ public class HeptaPrismController {
 
 	@FXML
 	private void calculate() {
-		int length = Integer.parseInt(baseSideTextField.getText());
-		int height = Integer.parseInt(heightTextField.getText());
+		double length = Double.parseDouble(baseSideTextField.getText());
+		double height = Double.parseDouble(heightTextField.getText());
 		HeptagonalPrism hepta = new HeptagonalPrism(height, length);
 		baseAreaTextField.setText(String.format("%.2f", hepta.getBaseArea()));
 		volumeTextField.setText(String.format("%.2f", hepta.getVolume()));
