+++ src/application/TetraPrismController.java
@@ -41,8 +41,8 @@ public class TetraPrismController {
 
 	@FXML
 	private void calculate() {
-		int length = Integer.parseInt(baseSideTextField.getText());
-		int height = Integer.parseInt(heightTextField.getText());
+		double length = Double.parseDouble(baseSideTextField.getText());
+		double height = Double.parseDouble(heightTextField.getText());
 		SquarePrism tetra = new SquarePrism(height, length);
 		baseAreaTextField.setText(String.format("%.2f", tetra.getBaseArea()));
 		volumeTextField.setText(String.format("%.2f", tetra.getVolume()));
